Change move generation from commands to use `(move:...)` links Fixes #34
